I've have also finished the Expressions group. This is the overall group activity that has happened here. Also
[16:37:02.178] The character encoding of the HTML document was not declared. The document will render with garbled text in some browser configurations if the document contains characters from outside the US-ASCII range. The character encoding of the page must be declared in the document or in the transfer protocol. @ file:///Users/bigdsweetz/Documents/FullSail/Web%20Programming%20Fundamentals/Week%201/First%20Git%20Project/Associating%20JS/index.html

This was caused in Firefox whenever I would try and run my code for the class lessons. I learned that Firefox need one more line of code before it will run with no problems

I searched through Google and came across an answer for my problem at https://github.com/webpack/webpack-dev-server/issues/1. The answer is to input a meta charset that will solve this issue.

For whatever reason, Firefox has not fixed this and they are currently working on resolving this. IT has been an ongoing issue since 2010.